removes tests above 7
